Remove duplicated itc stitch script and require passing a remote endopoint
