--- itcStitchOdbSchema.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-npm ci
-
-node itc/schemastitcher.mjs modules/schema/src/main/resources/lucuma/odb/graphql/OdbSchema.graphql itc/service/src/main/resources/graphql/itc_base.graphql itc/service/src/main/resources/graphql/itc.graphql
